requirements.txt
discord.py>=2.3.0
python-dotenv>=0.21.0
requests>=2.28.0